@@ -21,15 +21,26 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class CleanCommand extends Command
 {
+    /**
+     * @var TokenManagerInterface
+     */
     private $accessTokenManager;
+
+    /**
+     * @var TokenManagerInterface
+     */
     private $refreshTokenManager;
+
+    /**
+     * @var AuthCodeManagerInterface
+     */
     private $authCodeManager;
 
     public function __construct(
         TokenManagerInterface $accessTokenManager,
         TokenManagerInterface $refreshTokenManager,
-        AuthCodeManagerInterface $authCodeManager)
-    {
+        AuthCodeManagerInterface $authCodeManager
+    ) {
         parent::__construct();
 
         $this->accessTokenManager = $accessTokenManager;
@@ -40,14 +51,15 @@ class CleanCommand extends Command
     /**
      * {@inheritdoc}
      */
-    protected function configure()
+    protected function configure(): void
     {
         parent::configure();
 
         $this
             ->setName('fos:oauth-server:clean')
             ->setDescription('Clean expired tokens')
-            ->setHelp(<<<EOT
+            ->setHelp(
+                <<<EOT
 The <info>%command.name%</info> command will remove expired OAuth2 tokens.
 
   <info>php %command.full_name%</info>
@@ -59,11 +71,13 @@ EOT
     /**
      * {@inheritdoc}
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         foreach ([$this->accessTokenManager, $this->refreshTokenManager, $this->authCodeManager] as $service) {
             $result = $service->deleteExpired();
             $output->writeln(sprintf('Removed <info>%d</info> items from <comment>%s</comment> storage.', $result, get_class($service)));
         }
+
+        return 0;
     }
 }
