@@ -25,7 +25,6 @@ use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
@@ -47,11 +46,6 @@ class AuthorizeController
      * @var ClientInterface
      */
     private $client;
-
-    /**
-     * @var SessionInterface
-     */
-    private $session;
 
     /**
      * @var Form
@@ -103,17 +97,6 @@ class AuthorizeController
      * Thus, this is considered a bad practice to fetch services directly from container.
      *
      * @todo This controller could be refactored to not rely on so many dependencies
-     *
-     * @param RequestStack             $requestStack
-     * @param Form                     $authorizeForm
-     * @param AuthorizeFormHandler     $authorizeFormHandler
-     * @param OAuth2                   $oAuth2Server
-     * @param Environment              $twig
-     * @param TokenStorageInterface    $tokenStorage
-     * @param UrlGeneratorInterface    $router
-     * @param ClientManagerInterface   $clientManager
-     * @param EventDispatcherInterface $eventDispatcher
-     * @param SessionInterface|null    $session
      */
     public function __construct(
         RequestStack $requestStack,
@@ -124,11 +107,9 @@ class AuthorizeController
         TokenStorageInterface $tokenStorage,
         UrlGeneratorInterface $router,
         ClientManagerInterface $clientManager,
-        EventDispatcherInterface $eventDispatcher,
-        SessionInterface $session = null
+        EventDispatcherInterface $eventDispatcher
     ) {
         $this->requestStack = $requestStack;
-        $this->session = $session;
         $this->authorizeForm = $authorizeForm;
         $this->authorizeFormHandler = $authorizeFormHandler;
         $this->oAuth2Server = $oAuth2Server;
@@ -141,8 +122,7 @@ class AuthorizeController
 
     /**
      * Authorize.
-     * @param Request $request
-     * @return Response
+     *
      * @throws OAuth2RedirectException
      */
     public function authorizeAction(Request $request): Response
@@ -153,9 +133,9 @@ class AuthorizeController
             throw new AccessDeniedException('This user does not have access to this section.');
         }
 
-        if ($this->session && true === $this->session->get('_fos_oauth_server.ensure_logout')) {
-            $this->session->invalidate(600);
-            $this->session->set('_fos_oauth_server.ensure_logout', true);
+        if (!empty($request->getSession()) && true === $request->getSession()->get('_fos_oauth_server.ensure_logout')) {
+            $request->getSession()->invalidate(600);
+            $request->getSession()->set('_fos_oauth_server.ensure_logout', true);
         }
 
         $form = $this->authorizeForm;
@@ -168,7 +148,7 @@ class AuthorizeController
         );
 
         if ($event->isAuthorizedClient()) {
-            $scope = $request->get('scope' );
+            $scope = $request->get('scope');
 
             return $this->oAuth2Server->finishClientAuthorization(true, $user, $request, $scope);
         }
@@ -185,18 +165,11 @@ class AuthorizeController
         return $this->renderAuthorize($data);
     }
 
-    /**
-     * @param UserInterface        $user
-     * @param AuthorizeFormHandler $formHandler
-     * @param Request              $request
-     *
-     * @return Response
-     */
     protected function processSuccess(UserInterface $user, AuthorizeFormHandler $formHandler, Request $request): Response
     {
-        if ($this->session && true === $this->session->get('_fos_oauth_server.ensure_logout')) {
+        if (!empty($request->getSession()) && true === $request->getSession()->get('_fos_oauth_server.ensure_logout')) {
             $this->tokenStorage->setToken(null);
-            $this->session->invalidate();
+            $request->getSession()->invalidate();
         }
 
         $this->eventDispatcher->dispatch(
@@ -206,7 +179,7 @@ class AuthorizeController
 
         $formName = $this->authorizeForm->getName();
         if (!$request->query->all() && $request->request->has($formName)) {
-            $request->query->add($request->request->get($formName));
+            $request->query->add($request->request->all($formName));
         }
 
         try {
@@ -220,28 +193,19 @@ class AuthorizeController
 
     /**
      * Generate the redirection url when the authorize is completed.
-     *
-     * @param UserInterface $user
-     *
-     * @return string
      */
-    protected function getRedirectionUrl(UserInterface $user)
+    protected function getRedirectionUrl(UserInterface $user): string
     {
         return $this->router->generate('fos_oauth_server_profile_show');
     }
 
-    /**
-     * @return ClientInterface
-     */
-    protected function getClient()
+    protected function getClient(): ClientInterface
     {
         if (null !== $this->client) {
             return $this->client;
         }
 
-        if (null === $request = $this->getCurrentRequest()) {
-            throw new NotFoundHttpException('Client not found.');
-        }
+        $request = $this->getCurrentRequest();
 
         if (null === $clientId = $request->get('client_id')) {
             $formData = $request->get($this->authorizeForm->getName(), []);
@@ -258,8 +222,7 @@ class AuthorizeController
     }
 
     /**
-     * @param array $data
-     * @return Response
+     * @param array<string , mixed> $data Various data to be passed to the twig template
      * @throws LoaderError
      * @throws RuntimeError
      * @throws SyntaxError
@@ -271,13 +234,10 @@ class AuthorizeController
             $data
         );
 
-        return $response instanceof Response ? $response : new Response($response);
+        return new Response($response);
     }
 
-    /**
-     * @return Request|null
-     */
-    private function getCurrentRequest()
+    private function getCurrentRequest(): Request
     {
         $request = $this->requestStack->getCurrentRequest();
         if (null === $request) {
