@@ -13,8 +13,10 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\DependencyInjection\Security\Factory;
 
-use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\SecurityFactoryInterface;
+use Symfony\Bundle\SecurityBundle\DependencyInjection\Security\Factory\AuthenticatorFactoryInterface;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
+use Symfony\Component\DependencyInjection\Argument\ServiceClosureArgument;
+use Symfony\Component\DependencyInjection\ChildDefinition;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 
@@ -23,45 +25,59 @@ use Symfony\Component\DependencyInjection\Reference;
  *
  * @author Arnaud Le Blanc <arnaud.lb@gmail.com>
  */
-class OAuthFactory implements SecurityFactoryInterface
+class OAuthFactory implements  AuthenticatorFactoryInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function create(ContainerBuilder $container, $id, $config, $userProvider, $defaultEntryPoint)
+    public function createAuthenticator(ContainerBuilder $container, string $firewallName, array $config, string $userProviderId): string
     {
-        // NOTE: done like this to avoid PHPStan complaining about a missing class for both Symfony v3 and Symfony v4
-        $definitionDecorator = 'Symfony\\Component\\DependencyInjection\\DefinitionDecorator';
-        $childDefinition = 'Symfony\\Component\\DependencyInjection\\ChildDefinition';
-        $definitionClass = $childDefinition;
-        if (class_exists($definitionDecorator)) {
-            $definitionClass = $definitionDecorator;
-        }
+        $authenticatorId = 'security.authenticator.oauth2.'.$firewallName;
+        $firewallEventDispatcherId = 'security.event_dispatcher.'.$firewallName;
 
-        $providerId = 'security.authentication.provider.fos_oauth_server.'.$id;
+        // authenticator manager
+        $firewallAuthenticationProviders = [];
+        $authenticators = array_map(function ($firewallName) {
+            return new Reference($firewallName);
+        }, $firewallAuthenticationProviders);
         $container
-            ->setDefinition($providerId, new $definitionClass('fos_oauth_server.security.authentication.provider'))
-            ->replaceArgument(0, new Reference($userProvider))
+            ->setDefinition($managerId = 'security.authenticator.oauth2.'.$firewallName, new ChildDefinition('fos_oauth_server.security.authenticator.manager'))
+//            ->replaceArgument(0, $authenticators)
+//            ->replaceArgument(2, new Reference($firewallEventDispatcherId))
+            ->addTag('monolog.logger', ['channel' => 'security'])
         ;
 
-        $listenerId = 'security.authentication.listener.fos_oauth_server.'.$id;
-        $container->setDefinition($listenerId, new $definitionClass('fos_oauth_server.security.authentication.listener'));
+        $managerLocator = $container->getDefinition('security.authenticator.managers_locator');
+        $managerLocator->replaceArgument(0, array_merge($managerLocator->getArgument(0), [$firewallName => new ServiceClosureArgument(new Reference($managerId))]));
 
-        return [$providerId, $listenerId, 'fos_oauth_server.security.entry_point'];
+        // authenticator manager listener
+        $container
+            ->setDefinition('security.firewall.authenticator.'.$firewallName, new ChildDefinition('security.firewall.authenticator'))
+            ->replaceArgument(0, new Reference($managerId))
+        ;
+
+        // user checker listener
+        $container
+            ->setDefinition('security.listener.user_checker.'.$firewallName, new ChildDefinition('security.listener.user_checker'))
+            ->replaceArgument(0, new Reference('security.user_checker.'.$firewallName))
+            ->addTag('kernel.event_subscriber', ['dispatcher' => $firewallEventDispatcherId])
+        ;
+
+        // Add authenticators to the debug:firewall command
+        if ($container->hasDefinition('security.command.debug_firewall')) {
+            $debugCommand = $container->getDefinition('security.command.debug_firewall');
+            $debugCommand->replaceArgument(3, array_merge($debugCommand->getArgument(3), [$firewallName => $authenticators]));
+        }
+
+        return $authenticatorId;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getPosition()
+    public function getPriority(): int
     {
-        return 'pre_auth';
+        return 0;
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getKey()
+    public function getKey(): string
     {
         return 'fos_oauth';
     }
