@@ -18,38 +18,17 @@ use Symfony\Component\Security\Core\User\UserInterface;
 
 interface TokenInterface extends IOAuth2Token
 {
-    /**
-     * @param int $timestamp
-     */
-    public function setExpiresAt($timestamp);
+    public function setExpiresAt(int $timestamp);
 
-    /**
-     * @return int
-     */
-    public function getExpiresAt();
+    public function getExpiresAt(): int;
 
-    /**
-     * @param string $token
-     */
-    public function setToken($token);
+    public function setToken(string $token);
 
-    /**
-     * @param string $scope
-     */
-    public function setScope($scope);
+    public function setScope(string $scope);
 
-    /**
-     * @param UserInterface $user
-     */
     public function setUser(UserInterface $user);
 
-    /**
-     * @return UserInterface
-     */
-    public function getUser();
+    public function getUser(): ?UserInterface;
 
-    /**
-     * @param ClientInterface $client
-     */
     public function setClient(ClientInterface $client);
 }
