@@ -17,54 +17,36 @@ interface TokenManagerInterface
 {
     /**
      * Create a new TokenInterface.
-     *
-     * @return TokenInterface
      */
-    public function createToken();
+    public function createToken(): TokenInterface;
 
     /**
      * Return the class name of the Token.
-     *
-     * @return string
      */
-    public function getClass();
+    public function getClass(): string;
 
     /**
      * Retrieve a token using a set of criteria.
-     *
-     * @param array $criteria
-     *
-     * @return TokenInterface|null
      */
-    public function findTokenBy(array $criteria);
+    public function findTokenBy(array $criteria): ?TokenInterface;
 
     /**
      * Retrieve a token (object) by its token string.
-     *
-     * @param string $token a token
-     *
-     * @return TokenInterface|null
      */
-    public function findTokenByToken($token);
+    public function findTokenByToken(string $token): ?TokenInterface;
 
     /**
      * Save or update a given token.
-     *
-     * @param TokenInterface $token the token to save or update
      */
     public function updateToken(TokenInterface $token);
 
     /**
      * Delete a given token.
-     *
-     * @param TokenInterface $token the token to delete
      */
     public function deleteToken(TokenInterface $token);
 
     /**
      * Delete expired tokens.
-     *
-     * @return int the number of tokens deleted
      */
-    public function deleteExpired();
+    public function deleteExpired(): int;
 }
