@@ -28,7 +28,7 @@ class AuthorizeFormTypeTest extends TypeTestCase
      */
     protected $instance;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -42,7 +42,7 @@ class AuthorizeFormTypeTest extends TypeTestCase
         $this->instance = new AuthorizeFormType();
     }
 
-    public function testSubmit()
+    public function testSubmit(): void
     {
         $accepted = true;
         $formData = [
@@ -71,7 +71,7 @@ class AuthorizeFormTypeTest extends TypeTestCase
         }
     }
 
-    public function testConfigureOptionsWillSetDefaultsOnTheOptionsResolver()
+    public function testConfigureOptionsWillSetDefaultsOnTheOptionsResolver(): void
     {
         /** @var \PHPUnit_Framework_MockObject_MockObject|OptionsResolver $resolver */
         $resolver = $this->getMockBuilder(OptionsResolver::class)
@@ -91,19 +91,19 @@ class AuthorizeFormTypeTest extends TypeTestCase
         $this->assertNull($this->instance->configureOptions($resolver));
     }
 
-    public function testGetName()
+    public function testGetName(): void
     {
         $this->assertSame('fos_oauth_server_authorize', $this->instance->getName());
     }
 
-    public function testGetBlockPrefix()
+    public function testGetBlockPrefix(): void
     {
         $this->assertSame('fos_oauth_server_authorize', $this->instance->getBlockPrefix());
     }
 
     protected function getTypes()
     {
-        return  [
+        return [
             new AuthorizeFormType(),
         ];
     }
