@@ -13,35 +13,41 @@ declare(strict_types=1);
 
 namespace FOS\OAuthServerBundle\Tests\Functional;
 
+use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
+use FOS\OAuthServerBundle\FOSOAuthServerBundle;
+use FOS\OAuthServerBundle\Tests\Functional\TestBundle\TestBundle;
+use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
+use Symfony\Bundle\SecurityBundle\SecurityBundle;
+use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
 use Symfony\Component\HttpKernel\Kernel;
 
 class AppKernel extends Kernel
 {
-    public function registerBundles()
+    public function registerBundles(): array
     {
         $bundles = [
-            new \Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
-            new \Symfony\Bundle\SecurityBundle\SecurityBundle(),
-            new \Symfony\Bundle\TwigBundle\TwigBundle(),
-            new \FOS\OAuthServerBundle\FOSOAuthServerBundle(),
+            new FrameworkBundle(),
+            new SecurityBundle(),
+            new TwigBundle(),
+            new FOSOAuthServerBundle(),
 
-            new \FOS\OAuthServerBundle\Tests\Functional\TestBundle\TestBundle(),
+            new TestBundle(),
         ];
 
         if ('orm' === $this->getEnvironment()) {
-            $bundles[] = new \Doctrine\Bundle\DoctrineBundle\DoctrineBundle();
+            $bundles[] = new DoctrineBundle();
         }
 
         return $bundles;
     }
 
-    public function getCacheDir()
+    public function getCacheDir(): string
     {
         return sys_get_temp_dir().'/FOSOAuthServerBundle/';
     }
 
-    public function registerContainerConfiguration(LoaderInterface $loader)
+    public function registerContainerConfiguration(LoaderInterface $loader): void
     {
         $loader->load(__DIR__.'/config/config_'.$this->getEnvironment().'.yml');
     }
