@@ -22,6 +22,8 @@ use Symfony\Component\Security\Core\User\UserInterface;
 class User implements UserInterface
 {
     /**
+     * @var int
+     *
      * @ORM\Id
      * @ORM\Column(type="integer")
      * @ORM\GeneratedValue(strategy="AUTO")
@@ -29,40 +31,48 @@ class User implements UserInterface
     protected $id;
 
     /**
+     * @var string
+     *
      * @ORM\Column(type="string")
      */
     protected $password;
 
-    public function getId()
+    public function getId(): int
     {
         return $this->id;
     }
 
-    public function getRoles()
+    public function getRoles(): array
     {
         return ['ROLE_USER'];
     }
 
-    public function getPassword()
+    public function getPassword(): string
     {
         return $this->password;
     }
 
-    public function setPassword($password)
+    public function setPassword($password): void
     {
         $this->password = $password;
     }
 
-    public function getSalt()
+    public function getSalt(): ?string
     {
+        return null;
     }
 
-    public function getUsername()
+    public function getUsername(): string
     {
-        return $this->getId();
+        return (string)$this->getId();
     }
 
-    public function eraseCredentials()
+    public function getUserIdentifier(): string
+    {
+        return (string)$this->getId();
+    }
+
+    public function eraseCredentials(): void
     {
     }
 }
