@@ -11,26 +11,27 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FOS\OAuthServerBundle\Tests\Security\Authentification\Token;
+namespace FOS\OAuthServerBundle\Tests\Security\Authentication\Token;
 
 use FOS\OAuthServerBundle\Model\TokenInterface;
 use FOS\OAuthServerBundle\Security\Authentication\Token\OAuthToken;
+use PHPUnit\Framework\TestCase;
 
-class OAuthTokenTest extends \PHPUnit\Framework\TestCase
+class OAuthTokenTest extends TestCase
 {
     /**
      * @var OAuthToken
      */
     protected $instance;
 
-    public function setUp()
+    public function setUp(): void
     {
         $this->instance = new OAuthToken();
 
         parent::setUp();
     }
 
-    public function testSetTokenWillSetToken()
+    public function testSetTokenWillSetToken(): void
     {
         $token = $this->getMockBuilder(TokenInterface::class)
             ->disableOriginalConstructor()
@@ -41,7 +42,7 @@ class OAuthTokenTest extends \PHPUnit\Framework\TestCase
         $this->assertAttributeSame($token, 'token', $this->instance);
     }
 
-    public function testGetTokenWillReturnToken()
+    public function testGetTokenWillReturnToken(): void
     {
         $token = $this->getMockBuilder(TokenInterface::class)
             ->disableOriginalConstructor()
@@ -53,7 +54,7 @@ class OAuthTokenTest extends \PHPUnit\Framework\TestCase
         $this->assertSame($token, $this->instance->getToken());
     }
 
-    public function testGetCredentialsWillReturnToken()
+    public function testGetCredentialsWillReturnToken(): void
     {
         $token = $this->getMockBuilder(TokenInterface::class)
             ->disableOriginalConstructor()
